+++ app/Http/Controllers/apiController.php
@@ -20,27 +20,50 @@ class apiController extends Controller
     public function store(Request $request){ 
 
         $count = Rfid::count(); //check total record number
+        $allowAddTag = false;
               
         if (isset($_POST['Tag_id']) && isset($_POST['Tag_pc'])&&
         isset($_POST['Tag_count']) && isset($_POST['Tag_rssi'])&&
         isset($_POST['Reader_id']) && isset($_POST['Reader_record_time'])&&
         isset($_POST['Reader_ip'])) {
                 //debug use
-                var_dump($_POST['Tag_id']);
-                var_dump($_POST['Tag_pc']);
-                var_dump($_POST['Tag_count']);
-                var_dump($_POST['Tag_rssi']);
-                var_dump($_POST['Reader_id']);
-                var_dump($_POST['Reader_record_time']);
-                var_dump($_POST['Reader_ip']);
+                // var_dump($_POST['Tag_id']);
+                // var_dump($_POST['Tag_pc']);
+                // var_dump($_POST['Tag_count']);
+                // var_dump($_POST['Tag_rssi']);
+                // var_dump($_POST['Reader_id']);
+                // var_dump($_POST['Reader_record_time']);
+                // var_dump($_POST['Reader_ip']);
+                $rssi = 0;
+                if($_POST['Tag_rssi'] == "-52"){
+                    $rssi = -50;
+                }
+                else if($_POST['Tag_rssi'] == "-54"){
+                    $rssi = -52;
+                }
+                else if($_POST['Tag_rssi'] == "-58"){
+                    $rssi = -55;
+                }
+                else if($_POST['Tag_rssi'] == "-59"){
+                    $rssi = -57;
+                }
+                else if($_POST['Tag_rssi'] == "-64"){
+                    $rssi = -58;
+                }
+                else{
+                    $rssi = $_POST['Tag_rssi'];
+                }
+                
 
+                if($allowAddTag){
                 //For add a new tag to the all testing tag table
-                $record_check =  Alltag::where('tag_id','=',$_POST['Tag_id'])->first();
-                if($record_check == null){                 
-                    $Tag_id = $_POST['Tag_id'];
-                    Alltag::create([
-                        'tag_id' => $Tag_id
-                    ]);                   
+                    $record_check =  Alltag::where('tag_id','=',substr($_POST['Tag_id'], 1))->first();
+                    if($record_check == null){                 
+                        $Tag_id = substr($_POST['Tag_id'], 1);
+                        Alltag::create([
+                            'tag_id' => $Tag_id
+                        ]);                   
+                    }
                 }
 
                 // $losttime = new DateTime;
@@ -53,10 +76,12 @@ class apiController extends Controller
                 //                         ->where('updated_at','<=',$losttime)->get();  
                 
                 
+                $resetRecord = true; //clean the table 
+                $Trackstart = true; //important!!
 
                 //$resetRecord = false; //keep the table 
-
-                $resetRecord = true; //clean the table 
+                //$Trackstart = false;
+               
                 
                 if($resetRecord){
                     //Clean the data within 1 minutes
@@ -65,11 +90,13 @@ class apiController extends Controller
                     $Oneminformatted = $Onemin->format('Y-m-d H:i:s');
                     Onemins::where('created_at', '<=', $Oneminformatted)->delete();
 
-                    //Clean the data within 5 minutes
-                    $Fivemin = new DateTime;
-                    $Fivemin->modify('-5 minute');
-                    $Fiveminformatted = $Fivemin->format('Y-m-d H:i:s');
-                    Fivemins::where('created_at', '<=', $Fiveminformatted)->delete();
+                    if($Trackstart){
+                        //Clean the data within 5 minutes
+                        $Fivemin = new DateTime;
+                        $Fivemin->modify('-5 minute');
+                        $Fiveminformatted = $Fivemin->format('Y-m-d H:i:s');
+                        Fivemins::where('created_at', '<=', $Fiveminformatted)->delete();
+                    }
 
                     //Clean the data within 15 minutes
                     $Fifteenmin = new DateTime;
@@ -91,112 +118,82 @@ class apiController extends Controller
                 }
                 
             if($resetRecord){
+              
+                    $fakedata = false;
+                    $currentDate = date('Y-m-d H:i:s.') . gettimeofday()['usec'];
+                    if (strtotime($currentDate) - strtotime($_POST['Reader_record_time']) >= 5) {
+                         var_dump("Fake Data!!");
+                        var_dump($_POST['Tag_id']);
+                        var_dump($_POST['Tag_pc']);
+                        var_dump($_POST['Tag_count']);
+                        var_dump($rssi);
+                        var_dump($_POST['Reader_id']);
+                        var_dump($_POST['Reader_record_time']);
+                        var_dump($_POST['Reader_ip']);
+                        $fakedata = true;   
+                    }
+                    else{
+                        var_dump("True Data!!");
+                        $fakedata = false;
+                    } 
 
-                if($_POST['Tag_rssi'] != "-64"){
-                    Onemins::create([
-                        'tag_id' => $_POST['Tag_id'],
-                        'tag_rssi' => $_POST['Tag_rssi'],
-                        'tag_count' => $_POST['Tag_count'],                                               
-                        'reader_record_time' => $_POST['Reader_record_time'],
-                        'reader_ip' => $_POST['Reader_ip'],
-                        'created_at' => date('Y-m-d H:i:s'),
-                        'updated_at' => date('Y-m-d H:i:s')
-                    ]);
+                    if($fakedata == false){
+                        
+                            Onemins::create([
+                                'tag_id' => substr($_POST['Tag_id'], 1),
+                                'tag_rssi' => $rssi,
+                                'tag_count' => $_POST['Tag_count'],                                               
+                                'reader_record_time' => $_POST['Reader_record_time'],
+                                'reader_ip' => $_POST['Reader_ip'],
+                                'created_at' => date('Y-m-d H:i:s'),
+                                'updated_at' => date('Y-m-d H:i:s')
+                            ]);
 
-                    Fivemins::create([
-                        'tag_id' => $_POST['Tag_id'],
-                        'tag_rssi' => $_POST['Tag_rssi'],
-                        'tag_count' => $_POST['Tag_count'],                                               
-                        'reader_record_time' => $_POST['Reader_record_time'],
-                        'reader_ip' => $_POST['Reader_ip'],
-                        'created_at' => date('Y-m-d H:i:s'),
-                        'updated_at' => date('Y-m-d H:i:s')
-                    ]);
+                            if($Trackstart){                        
+                                Fivemins::create([
+                                    'tag_id' => substr($_POST['Tag_id'], 1),
+                                    'tag_rssi' => $rssi,
+                                    'tag_count' => $_POST['Tag_count'],                                               
+                                    'reader_record_time' => $_POST['Reader_record_time'],
+                                    'reader_ip' => $_POST['Reader_ip'],
+                                    'created_at' => date('Y-m-d H:i:s'),
+                                    'updated_at' => date('Y-m-d H:i:s')
+                                ]);
+                            }
 
-                    Fifteenmins::create([
-                        'tag_id' => $_POST['Tag_id'],
-                        'tag_rssi' => $_POST['Tag_rssi'],
-                        'tag_count' => $_POST['Tag_count'],                                               
-                        'reader_record_time' => $_POST['Reader_record_time'],
-                        'reader_ip' => $_POST['Reader_ip'],
-                        'created_at' => date('Y-m-d H:i:s'),
-                        'updated_at' => date('Y-m-d H:i:s')
-                    ]);
+                            Fifteenmins::create([
+                                'tag_id' => substr($_POST['Tag_id'], 1),
+                                'tag_rssi' => $rssi,
+                                'tag_count' => $_POST['Tag_count'],                                               
+                                'reader_record_time' => $_POST['Reader_record_time'],
+                                'reader_ip' => $_POST['Reader_ip'],
+                                'created_at' => date('Y-m-d H:i:s'),
+                                'updated_at' => date('Y-m-d H:i:s')
+                            ]);
 
-                    Halfhour::create([
-                        'tag_id' => $_POST['Tag_id'],
-                        'tag_rssi' => $_POST['Tag_rssi'],
-                        'tag_count' => $_POST['Tag_count'],                                               
-                        'reader_record_time' => $_POST['Reader_record_time'],
-                        'reader_ip' => $_POST['Reader_ip'],
-                        'created_at' => date('Y-m-d H:i:s'),
-                        'updated_at' => date('Y-m-d H:i:s')
-                    ]);
+                            Halfhour::create([
+                                'tag_id' => substr($_POST['Tag_id'], 1),
+                                'tag_rssi' => $rssi,
+                                'tag_count' => $_POST['Tag_count'],                                               
+                                'reader_record_time' => $_POST['Reader_record_time'],
+                                'reader_ip' => $_POST['Reader_ip'],
+                                'created_at' => date('Y-m-d H:i:s'),
+                                'updated_at' => date('Y-m-d H:i:s')
+                            ]);
 
-                    Onehour::create([
-                        'tag_id' => $_POST['Tag_id'],
-                        'tag_rssi' => $_POST['Tag_rssi'],
-                        'tag_count' => $_POST['Tag_count'],                                               
-                        'reader_record_time' => $_POST['Reader_record_time'],
-                        'reader_ip' => $_POST['Reader_ip'],
-                        'created_at' => date('Y-m-d H:i:s'),
-                        'updated_at' => date('Y-m-d H:i:s')
-                    ]); 
-
-
+                            Onehour::create([
+                                'tag_id' => substr($_POST['Tag_id'], 1),
+                                'tag_rssi' => $rssi,
+                                'tag_count' => $_POST['Tag_count'],                                               
+                                'reader_record_time' => $_POST['Reader_record_time'],
+                                'reader_ip' => $_POST['Reader_ip'],
+                                'created_at' => date('Y-m-d H:i:s'),
+                                'updated_at' => date('Y-m-d H:i:s')
+                            ]); 
+                      
+                    }
                 }
-                else{
-                    Onemins::create([
-                        'tag_id' => $_POST['Tag_id'],
-                        'tag_rssi' => -59.9,
-                        'tag_count' => $_POST['Tag_count'],                                               
-                        'reader_record_time' => $_POST['Reader_record_time'],
-                        'reader_ip' => $_POST['Reader_ip'],
-                        'created_at' => date('Y-m-d H:i:s'),
-                        'updated_at' => date('Y-m-d H:i:s')
-                    ]);
-
-                    Fivemins::create([
-                        'tag_id' => $_POST['Tag_id'],
-                        'tag_rssi' => -59.9,
-                        'tag_count' => $_POST['Tag_count'],                                               
-                        'reader_record_time' => $_POST['Reader_record_time'],
-                        'reader_ip' => $_POST['Reader_ip'],
-                        'created_at' => date('Y-m-d H:i:s'),
-                        'updated_at' => date('Y-m-d H:i:s')
-                    ]);
-
-                    Fifteenmins::create([
-                        'tag_id' => $_POST['Tag_id'],
-                        'tag_rssi' => -59.9,
-                        'tag_count' => $_POST['Tag_count'],                                               
-                        'reader_record_time' => $_POST['Reader_record_time'],
-                        'reader_ip' => $_POST['Reader_ip'],
-                        'created_at' => date('Y-m-d H:i:s'),
-                        'updated_at' => date('Y-m-d H:i:s')
-                    ]);
-
-                    Halfhour::create([
-                        'tag_id' => $_POST['Tag_id'],
-                        'tag_rssi' => -59.9,
-                        'tag_count' => $_POST['Tag_count'],                                               
-                        'reader_record_time' => $_POST['Reader_record_time'],
-                        'reader_ip' => $_POST['Reader_ip'],
-                        'created_at' => date('Y-m-d H:i:s'),
-                        'updated_at' => date('Y-m-d H:i:s')
-                    ]);
-
-                    Onehour::create([
-                        'tag_id' => $_POST['Tag_id'],
-                        'tag_rssi' => -59.9,
-                        'tag_count' => $_POST['Tag_count'],                                               
-                        'reader_record_time' => $_POST['Reader_record_time'],
-                        'reader_ip' => $_POST['Reader_ip'],
-                        'created_at' => date('Y-m-d H:i:s'),
-                        'updated_at' => date('Y-m-d H:i:s')
-                    ]); 
-                }
-            }               
+                          
                 
                 $record_check = Rfid::where('tag_id','=',substr($_POST['Tag_id'], 1))
                                     ->where('reader_ip','=',$_POST['Reader_ip'])
@@ -209,7 +206,7 @@ class apiController extends Controller
                         "tag_id"=>substr($_POST['Tag_id'], 1),
                         "tag_pc_value"=>$_POST['Tag_pc'],
                         "tag_count"=>$_POST['Tag_count'],
-                        "tag_rssi"=>$_POST['Tag_rssi'],
+                        "tag_rssi"=>$rssi,
                         "reader_antenna"=>$_POST['Reader_id'],
                         "reader_record_time"=>$_POST['Reader_record_time'],
                         "reader_ip"=>$_POST['Reader_ip'],
@@ -218,13 +215,12 @@ class apiController extends Controller
                     ->where('reader_ip','=',$_POST['Reader_ip'])
                     ->update($data);
                 }
-                else{       //a new record  
-                   
+                else{       //a new record                     
                     Rfid::create([
                         "tag_id"=>substr($_POST['Tag_id'], 1),
                         'tag_pc_value' =>$_POST['Tag_pc'],
                         'tag_count' => $_POST['Tag_count'],
-                        'tag_rssi' => $_POST['Tag_rssi'],
+                        'tag_rssi' => $rssi,
                         'reader_antenna' => $_POST['Reader_id'],
                         'reader_record_time' => $_POST['Reader_record_time'],
                         'reader_ip' => $_POST['Reader_ip'],
